Rename directive validation files
